Refactor namespaces in Product and GetProducts files
Updated namespaces in `ProductEndpoints.cs` and `GetProducts.cs` to reflect a focus on product management. Removed the import for `DigitalRightsManagement.Application.UserAggregate` in `ProductEndpoints.cs` and changed the namespace in `GetProducts.cs` from `UserAggregate` to `ProductAggregate`.
